feat: move all testimonials sections to reusable component
