--- src/components/pages/echo/testimonials/index.js
@@ -1,3 +1,0 @@
-import Testimonials from './testimonials';
-
-export default Testimonials;
