+++ src/components/shared/reusable-sections/testimonials/testimonials.jsx
@@ -13,7 +13,7 @@ const NextArrow = (props) => {
 
   return (
     <button
-      className="absolute right-0 top-[-74px] flex justify-center items-center w-11 h-11 border border-gray-5 hover:bg-gray-4 hover:border-gray-4 rounded-md lg:top-[-70px] md:w-9 md:h-9 md:-top-16 sm:top-[calc(100%+30px)] sm:right-[calc(100%-82px)] transition-colors duration-200"
+      className="absolute right-0 top-[-74px] flex h-11 w-11 items-center justify-center rounded-md border border-gray-5 transition-colors duration-200 hover:border-gray-4 hover:bg-gray-4 lg:top-[-70px] md:-top-16 md:h-9 md:w-9 sm:right-[calc(100%-82px)] sm:top-[calc(100%+30px)]"
       type="button"
       aria-label="Next testimonial"
       onClick={onClick}
@@ -31,7 +31,7 @@ const PrevArrow = (props) => {
 
   return (
     <button
-      className="absolute right-[58px] top-[-74px] flex justify-center items-center w-11 h-11 border border-gray-5 hover:bg-gray-4 hover:border-gray-4 rounded-md lg:top-[-70px] md:w-9 md:h-9 md:-top-16 md:right-[46px] sm:top-[calc(100%+30px)] sm:right-[calc(100%-36px)] transition-colors duration-200"
+      className="absolute right-[58px] top-[-74px] flex h-11 w-11 items-center justify-center rounded-md border border-gray-5 transition-colors duration-200 hover:border-gray-4 hover:bg-gray-4 lg:top-[-70px] md:-top-16 md:right-[46px] md:h-9 md:w-9 sm:right-[calc(100%-36px)] sm:top-[calc(100%+30px)]"
       type="button"
       aria-label="Prev testimonial"
       onClick={onClick}
@@ -55,31 +55,34 @@ const Testimonials = ({ title, testimonials }) => {
   };
 
   return (
-    <section className="testimonials safe-paddings mt-40 lg:mt-[120px] md:mt-[100px] sm:pb-16 sm:mt-20">
-      <div className="container-sm px-[80px] relative md:max-w-xl md:px-8 sm:w-full sm:px-5">
+    <section className="testimonials safe-paddings mt-40 lg:mt-[120px] md:mt-[100px] sm:mt-20 sm:pb-16">
+      <div className="container-sm relative px-[80px] md:max-w-xl md:px-8 sm:w-full sm:px-5">
         <Heading
-          className="text-4xl leading-tight tracking-snug font-medium lg:text-[32px] md:text-3xl"
+          className="relative z-10 pr-28 font-medium leading-denser tracking-snug lg:text-[32px] md:text-3xl"
           size="md"
           tag="h2"
         >
           {title}
         </Heading>
-        <Slider className="!flex w-full justify-between mt-8 md:mt-7 sm:mt-[30px]" {...settings}>
+        <Slider className="relative z-10 mt-7 !flex w-full justify-between sm:mt-6" {...settings}>
           {testimonials.map(({ content, avatar, name, company }, index) => (
-            <figure className="" key={index}>
+            <figure key={index}>
               <blockquote>
-                <p className="text-[22px] font-book tracking-[-0.04em] text-transparent bg-clip-text bg-[linear-gradient(98.67deg,rgba(255,255,255,0.85)_9.63%,rgba(255,255,255,0.6)_113.79%)] lg:text-xl md:text-lg sm:text-[15px]">
+                <p className="bg-testimonials-text bg-clip-text text-xl font-book tracking-tighter text-transparent md:text-lg sm:text-base">
                   {content}
                 </p>
               </blockquote>
 
               <figcaption className="mt-3.5 sm:mt-4">
                 <div className="flex items-center">
-                  <div className="rounded-full w-8 h-8 overflow-hidden grayscale mr-3.5">
-                    {avatar}
-                  </div>
+                  {avatar && (
+                    <div className="mr-3.5 h-8 w-8 overflow-hidden rounded-full grayscale">
+                      {avatar}
+                    </div>
+                  )}
                   <p className="text-sm leading-snug">
-                    {name} {company && <span className="font-light ml-0.5">- {company}</span>}
+                    {name}
+                    {company && <span className="ml-0.5 font-light">{company}</span>}
                   </p>
                 </div>
               </figcaption>
@@ -87,7 +90,7 @@ const Testimonials = ({ title, testimonials }) => {
           ))}
         </Slider>
         <img
-          className="absolute top-[30px] -left-10 sm:hidden lg:top-6 md:w-[116px] md:h-auto md:left-[-65px]"
+          className="pointer-events-none absolute -left-10 top-[30px] select-none lg:top-6 md:left-[-65px] md:h-auto md:w-[116px] sm:hidden"
           src={blockquote}
           alt=""
           width={144}
@@ -96,7 +99,7 @@ const Testimonials = ({ title, testimonials }) => {
           aria-hidden
         />
         <div
-          className="absolute -left-16 -top-9 w-[278px] h-[267px] opacity-[0.16] bg-[radial-gradient(65.61%_130.08%_at_74.29%_61.64%,#B7FBFF_0%,#96C6FF_46.5%,#4791FF_100%)] rounded-full blur-3xl transform-gpu md:-left-32"
+          className="absolute -left-16 -top-9 h-[267px] w-[278px] transform-gpu rounded-full bg-[radial-gradient(65.61%_130.08%_at_74.29%_61.64%,#B7FBFF_0%,#96C6FF_46.5%,#4791FF_100%)] opacity-[0.16] blur-3xl md:-left-32"
           aria-hidden
         />
       </div>
